holding position was added

+++ modules/drawer.js
@@ -1,3 +1,4 @@
+import { DragDrop } from './dragNdrop.js';
 import { Storage } from './storage.js';
 
 class Task {
@@ -5,6 +6,7 @@ class Task {
     this.description = description;
     this.status = 'todo';
     this.uid = +new Date();
+    this.count = Storage.get(this.status).length;
   }
 }
 export class Drawer {
@@ -30,7 +32,7 @@ export class Drawer {
 
   static drawTasks() {
     this.columns.forEach((col) => {
-      console.log(`-----------Список ${col}------------`);
+       console.log(`-----------Список ${col}------------`);
       let column = document.getElementById(col);
       column.innerHTML = '';
       const tasks = Storage.get(col);
@@ -55,21 +57,21 @@ export class Drawer {
   }
 
   static finish(dragElement, currentDroppable) {
+    const oldStatus = dragElement.getAttribute('data-status');
+    let drawn = false;
     if (!currentDroppable) {
       this.drawTasks();
       return;
     }
     this.columns.forEach((col) => {
       if (currentDroppable.getAttribute('id') == dragElement.getAttribute('data-status')) {
+        if (drawn) return;
+        this.setPosition(dragElement, currentDroppable, oldStatus);
         this.drawTasks();
-        //return;
+        drawn = true;
       } else if (currentDroppable.getAttribute('id') == col) {
-        const oldStatus = dragElement.getAttribute('data-status');
-        const uid = +dragElement.getAttribute('uid');
-        const list = Storage.get(oldStatus);
-        const entity = list.find((item) => item.uid === uid);
-        entity.status = col;
-        this.change(entity, oldStatus);
+        this.setPosition(dragElement, currentDroppable, oldStatus);
+        this.drawTasks();
       } else if (currentDroppable.classList.contains('bin')) {
         const uid = +dragElement.getAttribute('uid');
         Storage.delete(dragElement.getAttribute('data-status'), uid);
@@ -78,18 +80,50 @@ export class Drawer {
     });
   }
 
-  static change(entity, oldLocation) {
-    const newLocation = entity.status;
+  static change(entity, oldLocation, newLocation) {
     Storage.add(newLocation, entity);
-    redefinition(newLocation);
+    this.redefinition(newLocation);
+    if (newLocation == oldLocation) return; // Посмотреть эту фичу
     Storage.delete(oldLocation, entity.uid);
-    redefinition(oldLocation);
+    this.redefinition(oldLocation);
     this.drawTasks();
-    function redefinition(location) {
-      Storage.get(location).forEach((item, index) => {
+  }
+
+  static redefinition(location) {
+    Storage.get(location)
+      .sort((a, b) => a.count - b.count)
+      .forEach((item, index) => {
         item.count = index;
         Storage.update(location, item);
       });
-    }
+  }
+
+  static setPosition(dragElement, currentDroppable, oldStatus) {
+    const positions = [];
+    const dragElementCoords = dragElement.getBoundingClientRect();
+    const dragElementMiddleY = dragElementCoords.top + dragElementCoords.height / 2;
+    const cards = currentDroppable.querySelectorAll('.draggable.card');
+    cards.forEach((item) => {
+      if (item == dragElement) return;
+      let coords = item.getBoundingClientRect();
+      let middleY = coords.top + coords.height / 2;
+      positions.push(middleY);
+    });
+
+    const newPos = positions.filter((top) => top < dragElementMiddleY).length;
+
+    const uid = +dragElement.getAttribute('uid');
+    const list = Storage.get(oldStatus);
+    const entity = list.find((item) => item.uid === uid);
+    entity.status = currentDroppable.getAttribute('id');
+    entity.count = newPos;
+    const tmp = { ...entity };
+    Storage.delete(oldStatus, entity.uid);
+
+    const newList = Storage.get(currentDroppable.id);
+    newList.sort((a, b) => a.count - b.count);
+    newList.splice(newPos, 0, tmp);
+    newList.forEach((item, index) => (item.count = index));
+    Storage.bulk(currentDroppable.id, newList)
   }
 }
